+++ figures/point.h
@@ -27,13 +27,13 @@ public:
      * \brief Установить значение по X
      * \param x
      */
-    void setX(int x);
+    void setX(double x);
 
     /*!
      * \brief Установить значение по Y
      * \param y
      */
-    void setY(int y);
+    void setY(double y);
 
     /*!
      * \brief Установить сразу две координаты
@@ -43,7 +43,14 @@ public:
     void setPosition(double x, double y);
 
 private:
+    /*!
+     * \brief Позиция в точке X
+     */
     double m_positionX;
+
+    /*!
+     * \brief Позиция в точке Y
+     */
     double m_positionY;
 };
